+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Current (in progress)
 
 - Front modifications to display resources schema url field [#220](https://github.com/etalab/udata-front/pull/220)
+- Update ventti button url [#223](https://github.com/etalab/udata-front/pull/223)
+- Dataset page links fix [#219](https://github.com/etalab/udata-front/pull/219)
 
 ## 3.2.0 (2023-03-07)
 
@@ -11,6 +13,7 @@
   - `contextfunction` and `contextfilter` from Jinja is deprecated and replaced by `pass_context`
   - Move `Flask-Themes2` dep from udata to udata-front and upgrade to 1.0.0
 - Fix `@background-contrast-blue-cumulus` LESS variable [#217](https://github.com/etalab/udata-front/pull/217)
+- Fix "informations" throughout the repo [#218](https://github.com/etalab/udata-front/pull/218)
 
 ## 3.1.3 (2023-03-02)