+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Current (in progress)
 
+- Update ventti button url [#223](https://github.com/etalab/udata-front/pull/223)
+- Dataset page links fix [#219](https://github.com/etalab/udata-front/pull/219)
 - Fix "informations" throughout the repo [#218](https://github.com/etalab/udata-front/pull/218)
 - fix RGAA criterion 9.3 [#222](https://github.com/etalab/udata-front/pull/222)