+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Update ventti button url [#223](https://github.com/etalab/udata-front/pull/223)
 - Dataset page links fix [#219](https://github.com/etalab/udata-front/pull/219)
+- Fix "informations" throughout the repo [#218](https://github.com/etalab/udata-front/pull/218)
+- fix RGAA criterion 9.3 [#222](https://github.com/etalab/udata-front/pull/222)
 
 ## 3.2.0 (2023-03-07)
 
@@ -12,7 +14,6 @@
   - `contextfunction` and `contextfilter` from Jinja is deprecated and replaced by `pass_context`
   - Move `Flask-Themes2` dep from udata to udata-front and upgrade to 1.0.0
 - Fix `@background-contrast-blue-cumulus` LESS variable [#217](https://github.com/etalab/udata-front/pull/217)
-- Fix "informations" throughout the repo [#218](https://github.com/etalab/udata-front/pull/218)
 
 ## 3.1.3 (2023-03-02)