+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Update ventti button url [#223](https://github.com/etalab/udata-front/pull/223)
 - Dataset page links fix [#219](https://github.com/etalab/udata-front/pull/219)
+- Fix "informations" throughout the repo [#218](https://github.com/etalab/udata-front/pull/218)
+- fix RGAA criterion 9.3 [#222](https://github.com/etalab/udata-front/pull/222)
 - New reuse page [#210](https://github.com/etalab/udata-front/pull/210)
 
 ## 3.2.0 (2023-03-07)
@@ -13,7 +15,6 @@
   - `contextfunction` and `contextfilter` from Jinja is deprecated and replaced by `pass_context`
   - Move `Flask-Themes2` dep from udata to udata-front and upgrade to 1.0.0
 - Fix `@background-contrast-blue-cumulus` LESS variable [#217](https://github.com/etalab/udata-front/pull/217)
-- Fix "informations" throughout the repo [#218](https://github.com/etalab/udata-front/pull/218)
 
 ## 3.1.3 (2023-03-02)
 
@@ -25,6 +26,7 @@
 - Show a warning notice when JavaScript is disabled or failed to execute [#206](https://github.com/etalab/udata-front/pull/206)
 - Update lighthouse to fix security issues [#211](https://github.com/etalab/udata-front/pull/211)
 - New dataset page [#181](https://github.com/etalab/udata-front/pull/181)
+- Update vue-toaster dependency to avoid reported vulnerabilities [#215](https://github.com/etalab/udata-front/pull/215)
 
 ## 3.1.2 (2023-02-06)