+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Front modifications to display resources schema url field [#220](https://github.com/etalab/udata-front/pull/220)
+- Update ventti button url [#223](https://github.com/etalab/udata-front/pull/223)
+- Dataset page fixes [#219](https://github.com/etalab/udata-front/pull/219) [#229](https://github.com/etalab/udata-front/pull/229)
+- Fix "informations" throughout the repo [#218](https://github.com/etalab/udata-front/pull/218)
+- fix RGAA criterion 9.3 [#222](https://github.com/etalab/udata-front/pull/222)
+- New reuse page [#210](https://github.com/etalab/udata-front/pull/210)
+- Fix [dependabot/10](https://github.com/etalab/udata-front/security/dependabot/10) [#228](https://github.com/etalab/udata-front/pull/228)
 
 ## 3.2.0 (2023-03-07)
 
@@ -11,7 +17,6 @@
   - `contextfunction` and `contextfilter` from Jinja is deprecated and replaced by `pass_context`
   - Move `Flask-Themes2` dep from udata to udata-front and upgrade to 1.0.0
 - Fix `@background-contrast-blue-cumulus` LESS variable [#217](https://github.com/etalab/udata-front/pull/217)
-- Fix "informations" throughout the repo [#218](https://github.com/etalab/udata-front/pull/218)
 
 ## 3.1.3 (2023-03-02)
 
@@ -23,10 +28,12 @@
 - Show a warning notice when JavaScript is disabled or failed to execute [#206](https://github.com/etalab/udata-front/pull/206)
 - Update lighthouse to fix security issues [#211](https://github.com/etalab/udata-front/pull/211)
 - New dataset page [#181](https://github.com/etalab/udata-front/pull/181)
+- Update vue-toaster dependency to avoid reported vulnerabilities [#215](https://github.com/etalab/udata-front/pull/215)
 
 ## 3.1.2 (2023-02-06)
 
 - Use brand color for button style `tertiary-no-outline` [#199](https://github.com/etalab/udata-front/pull/199)
+- Use computed dates for resources and datasets [#201](https://github.com/etalab/udata-front/pull/201)
 - Fix setuptools version used in CI [#202](https://github.com/etalab/udata-front/pull/202)
 - Move Pagination to `@etalab/udata-front-plugin-helpers` and add tests [#200](https://github.com/etalab/udata-front/pull/200)