+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Current (in progress)
+## 3.2.1 (2023-03-28)
 
 - Front modifications to display resources schema url field [#220](https://github.com/etalab/udata-front/pull/220)
 - Update ventti button url [#223](https://github.com/etalab/udata-front/pull/223)